### 2022-12-06 #4983 @kthyng
@bmcfee Would this be an ok fit for you as editor?
