### 2022-12-07 #4983 @ajakef
What makes a good JOSS reviewer? I saw the requirement to have a github account and have the software's requirements installed (Python). Is it best to suggest people that have used the software, or people who haven't used the software but potentially could in the future? Does geophysics/infrasound expertise matter?
