### 2022-12-08 #4983 @kthyng
@ajakef the reviewers install the software and go through the [reviewer checklist](https://joss.readthedocs.io/en/latest/review_checklist.html), opening up issues on the software repo as they come across them. It's nice to get reviewers who can focus on different aspects of the software: maybe one who is an expert with the subject and another who is more of an expert with the technique. If you find someone who has used the software but is not conflicted, that can be good because they have more specific ideas of what to look for in the software functionality. Otherwise it doesn't matter if they have used it before.
