### 2022-12-09 #4983 @ajakef
Thanks for that clarification. Here are a couple scientists who have github accounts, have used gemlog, and don't have a conflict of interest per https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy.
Wes Thelen, thelenwes, Cascades Volcano Observatory, USGS, wthelen@usgs.gov
Anna Perttu, aperttu, Massey University, A.Perttu@massey.ac.nz
