+++ paper/paper.bib
@@ -124,7 +124,7 @@ year = {in prep.}
   year={2020}
 }
 @techreport{Dannemann2022,
-  title={Data Report: TurboWave I and II Data Release.},
+  title={Data Report: TurboWave {I} and {{II}} Data Release.},
   author={Dannemann Dugick, Fransiska and Bowman, Daniel},
   year={2022},
   institution={Sandia National Lab.(SNL-NM), Albuquerque, NM (United States)}
