+++ gemlog/xcorr.py
@@ -1,0 +1,192 @@
+from obspy.signal.cross_correlation import correlate, xcorr_max
+import numpy as np
+import pandas as pd
+import glob, os
+from gemlog.gem_network import _unique
+
+xcorr_all(pattern = '*00..1*[0-3]', t1 = '2020-04-20', t2 = '2020-04-22')
+
+st = obspy.read('2020-04-20*00..10[0-3]*')
+output = xcorr_one_day(st)
+
+def xcorr_all(path = '.', pattern = '*', t1 = '1970-01-01', t2 = '9999-12-31', IDs = None):
+    return loop_through_days(xcorr_one_day, path, pattern, t1, t2, IDs)
+    
+
+
+def loop_through_days(function, path = '.', pattern = '*', t1 = '1970-01-01', t2 = '9999-12-31', IDs = None):
+    try:
+        t1 = obspy.UTCDateTime(t1)
+    except:
+        raise Exception('invalid t1')
+    try:
+        t2 = obspy.UTCDateTime(t2)
+    except:
+        raise Exception('invalid t2')
+    
+    ## make a database of files
+    file_metadata = {'filename':[], 't1':[], 't2':[], 'ID':[]}
+    filenames = glob.glob(os.path.join(path, pattern))
+    for filename in filenames:
+        st = obspy.read(filename, headonly=True)
+        for tr in st:
+            file_metadata['filename'].append(filename)
+            file_metadata['t1'].append(tr.stats.starttime)
+            file_metadata['t2'].append(tr.stats.endtime)
+            file_metadata['ID'].append(tr.id)
+
+    file_metadata_df = pd.DataFrame.from_dict(file_metadata)
+
+    t1 = np.max([t1, file_metadata_df.t1.min()])
+    t2 = np.min([t2, file_metadata_df.t2.max()])
+    if IDs is None:
+        IDs = _unique(file_metadata_df.ID)
+
+    rows_to_keep = np.where((file_metadata_df.t2 >= t1) & \
+                    (file_metadata_df.t1 <= t2) & \
+                    np.isin(file_metadata_df.ID, IDs))[0]
+    file_metadata_df = file_metadata_df.iloc[rows_to_keep,:]
+    file_metadata_df.sort_values('t1', ignore_index=True, inplace=True)
+
+    ## loop through days. be careful to avoid funny business with leap seconds.
+    day_starts = [t1]
+    while True:
+        new_time = (day_starts[-1]+86400+2).replace(hour=0, minute=0, second=0) # add 24 hours + 2 sec, then round down
+        if new_time < t2:
+            day_starts.append(new_time)
+        else:
+            break
+    day_starts = np.array(day_starts)
+
+    day_ends = [day_starts[1]]
+    while True:
+        new_time = (day_ends[-1]+86400+2).replace(hour=0, minute=0, second=0) # add 24 hours + 2 sec, then round down
+        if new_time < t2:
+            day_ends.append(new_time)
+        else:
+            day_ends.append(t2)
+            break
+    day_ends = np.array(day_ends)
+
+    files_read = []
+    output_list = []
+    st = obspy.Stream()
+    for (day_start, day_end) in zip(day_starts, day_ends):
+        indices = np.where((file_metadata_df.t1 <= day_end) &
+                           (file_metadata_df.t2 >= day_start))[0]
+        ## for efficiency, avoid reading files twice if they cover multiple days or IDs
+        for filename in file_metadata_df.filename[indices]:
+            if filename not in files_read:
+                st += obspy.read(filename)
+                files_read.append(filename)
+
+        st.merge()
+        ## throw out data before the start of this day, and any unused traces
+        st.trim(day_start, t2)
+        for tr in st:
+            if tr.id not in IDs:
+                st.pop(tr)
+
[…]
