+++ gemlog/xcorr.py
@@ -1,18 +1,61 @@
 from obspy.signal.cross_correlation import correlate, xcorr_max
+import obspy
 import numpy as np
 import pandas as pd
 import glob, os
 from gemlog.gem_network import _unique
 
-xcorr_all(pattern = '*00..1*[0-3]', t1 = '2020-04-20', t2 = '2020-04-22')
+#data = xcorr_all(pattern = '*00..1*[0-3]', t1 = '2020-04-20', t2 = '2020-04-22')
 
-st = obspy.read('2020-04-20*00..10[0-3]*')
-output = xcorr_one_day(st)
+#st = obspy.read('2020-04-20*00..10[0-3]*')
+#output = xcorr_one_day(st)
+
 
 def xcorr_all(path = '.', pattern = '*', t1 = '1970-01-01', t2 = '9999-12-31', IDs = None):
     return loop_through_days(xcorr_one_day, path, pattern, t1, t2, IDs)
-    
 
+def invert_for_slowness(xcorr_df, locations):
+    """
+    locations: pd.DataFrame with columns x, y, and network/station/location (output of get_coordinates)
+    """
+    data_short_IDs = []
+    for key in xcorr_df.keys():
+        if re.search('rms', key):
+            data_short_IDs.append(key.split('_')[1])
+            
+    locations['ID'] = [f'{locations.network[i]}.{locations.station[i]}.{locations.location[i]}' for i in range(locations.shape[0])]
+    keep_indices = np.where(np.isin(locations.ID, data_short_IDs))[0]
+    locations = locations.iloc[keep_indices, :]
+    locations.sort_values('ID', inplace=True, ignore_index = True)
+
+    ## solve linear system G . s = t: G is x/y distances, s is slowness, and t is observed time lags
+    G = []
+    lag_keys = []
+    for key in xcorr_df.keys():
+        if re.search('lag', key):
+            lag_keys.append(key)
+            (short_ID_1, short_ID_2) = key.split('_')[1:]
+            index1 = np.where(locations.ID == short_ID_1)[0][0]
+            index2 = np.where(locations.ID == short_ID_2)[0][0]
+            G.append([locations.loc[index1, 'x'] - locations.loc[index2, 'x'],
+                      locations.loc[index1, 'y'] - locations.loc[index2, 'y']])
+
+    G = np.array(G)[:-1,:] # drop the last row because it's linearly dependent on the others
+    
+    ## Use the generalized inverse in case there are more than 3 sensors: (GTG)^-1 . GT . t = s
+    ## @ is matrix multiplication symbol
+    G_inv_gen = np.linalg.inv(G.T @ G) @ G.T
+
+    ## loop through all the time windows and make a slowness vector for each
+    n_windows = xcorr_df.shape[0]
+    azimuth = np.zeros(n_windows)
+    slowness = np.zeros(n_windows)
+    for i in range(n_windows):
+        lags = np.array([xcorr_df[lag_key][i] for lag_key in lag_keys])
+        slowness_vector = G_inv_gen @ lags
+        azimuth[i] = np.atan2(slowness_vector[0], slowness_vector[1])
+        slowness[i] = np.sqrt(np.sum(slowness_vector**2))
+    
 
 def loop_through_days(function, path = '.', pattern = '*', t1 = '1970-01-01', t2 = '9999-12-31', IDs = None):
     try:
@@ -27,20 +70,25 @@ def loop_through_days(function, path = '.', pattern = '*', t1 = '1970-01-01', t2
     ## make a database of files
     file_metadata = {'filename':[], 't1':[], 't2':[], 'ID':[]}
     filenames = glob.glob(os.path.join(path, pattern))
+    if len(filenames) == 0:
[…]
