### 2023-03-14 04030be
progress on UI for xcorr
