### 2023-03-14 c814da1
bug fix in verify_huddle_test
