+++ gemlog/version.py
@@ -1,4 +1,5 @@
-__version__ = '1.6.6' # version described by JOSS paper 
+__version__ = '1.6.7' # bug fix in verify_huddle_test 
+#__version__ = '1.6.6' # version described by JOSS paper 
 #__version__ = '1.6.5' # bug fixes and other minor improvements 
 #__version__ = '1.6.4' # upgrades dependencies due to moderate security issue in numpy: https://github.com/advisories/GHSA-fpfv-jqm9-f5jm 
 #__version__ = '1.6.3' # minor improvements/bug fixes for verify_huddle_test 
@@ -58,4 +59,4 @@ __version__ = '1.6.6' # version described by JOSS paper
 #__version__  = '0.2.2' # automated github tests, setup.py improvements, and modelst speed-up
 #__version__  = '0.2.1' # 
 #__version__ = '0.0.5' # added new functions to make network map from gps data and rename mseeds from serial_number.channel to network.station.location.channel codes
-## List of old versions is not comprehensive##################################################
+## List of old versions is not comprehensive###################################################
