### 2023-03-14 #5256 @editorialbot
[REVIEW]: gemlog: Data Conversion for the Open-Source Gem Infrasound Logger
**Submitting author:** @ajakef (<a href="http://orcid.org/0000-0001-6447-6778">Jacob Anderson</a>)
**Repository:** https://github.com/ajakef/gemlog
**Branch with paper.md** (empty if default branch): paper
**Version:** 1.6.5.0
