+++ paper/paper.bib
@@ -6,7 +6,8 @@
   number={1},
   pages={153--164},
   year={2018},
-  publisher={GeoScienceWorld}
+  publisher={GeoScienceWorld},
+  doi = {10.1785/0220170067}
 }
 
 @article{Brissaud2021,
@@ -17,14 +18,16 @@
   number={12},
   pages={e2021GL093013},
   year={2021},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi = {10.1029/2021gl093013}
 }
 
 @article{Krishnamoorthy2020,
   title={Infrasound Measurements in the Arctic on a Long-Duration Balloon},
   author={Krishnamoorthy, Siddharth and Bowman, Daniel C and Komjathy, Attila and Pauken, Michael T and Cutts, James A and Carlo, Marine De and Ardhuin, Fabrice and Le Pichon, Alexis},
   year={2020},
-  publisher={Pasadena, CA: Jet Propulsion Laboratory, National Aeronautics and Space Administration}
+  publisher={Pasadena, CA: Jet Propulsion Laboratory, National Aeronautics and Space Administration},
+  url = {https://hdl.handle.net/2014/52499}
 }
 @article{Young2018,
   title={Explosion-generated infrasound recorded on ground and airborne microbarometers at regional distances},
@@ -34,13 +37,15 @@
   number={4},
   pages={1497--1506},
   year={2018},
-  publisher={GeoScienceWorld}
+  publisher={GeoScienceWorld},
+  doi = {10.1785/0220180038}
 }
 @techreport{SladMerchant2021,
   title={Evaluation of Low Cost Infrasound Sensor Packages.},
   author={Slad, George and Merchant, Bion},
   year={2021},
-  institution={Sandia National Lab.(SNL-NM), Albuquerque, NM (United States)}
+  institution={Sandia National Lab.(SNL-NM), Albuquerque, NM (United States)},
+  doi = {10.2172/1829264}
 }
 @article{Rosenblatt2022,
   title={Controls on the frequency content of near-source infrasound at open-vent volcanoes: a case study from {V}olc{\'a}n {V}illarrica, {C}hile},
@@ -50,7 +55,8 @@
   number={12},
   pages={1--17},
   year={2022},
-  publisher={Springer}
+  publisher={Springer},
+  doi = {10.1007/s00445-022-01607-y}
 }
 @article{BowmanAlbert2018,
   title={Acoustic event location and background noise characterization on a free flying infrasound sensor network in the stratosphere},
@@ -60,13 +66,15 @@
   number={3},
   pages={1524--1535},
   year={2018},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi = {10.1093/gji/ggy069}
 }
 @phdthesis{Ronan2017,
   title={Fluvial seismo-acoustics: Characterizing temporally dependent river conditions through near river seismic and acoustic signals},
   author={Ronan, Timothy J},
   year={2017},
-  school={The University of North Carolina at Chapel Hill}
+  school={The University of North Carolina at Chapel Hill},
+  doi = {10.17615/nseg-wy10}
 }
 @article{Averbuch2022,
   title={Evidence for Short Temporal Atmospheric Variations Observed by Infrasonic Signals: 1. The Troposphere},
@@ -76,7 +84,8 @@
   number={3},
   pages={e2021EA002036},
   year={2022},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi = {10.1029/2021ea002036}  
 }
 
 @article{Anderson_inprep,
@@ -112,7 +121,8 @@ year = {in prep.}
   number={1},
[…]
