### 2023-03-15 018e444
update scipy version in requirements-min.txt to prevent installation error
