+++ requirements-min.txt
@@ -1,5 +1,5 @@
 numpy==1.22
 pandas==1.3.0
-scipy==1.3.0
+scipy==1.4.0
 matplotlib==3.2.0
 obspy==1.3.0
