+++ gemlog/xcorr.py
@@ -2,34 +2,108 @@ from obspy.signal.cross_correlation import correlate, xcorr_max
 import obspy
 import numpy as np
 import pandas as pd
-import glob, os, traceback, sys, getopt, argparse
+import glob, os, traceback, sys, getopt, argparse, re
 from gemlog.gem_network import _unique
+import scipy.interpolate
 
-def main():
-    parser = argparse.ArgumentParser(description='Use cross-correlation to find delays between waveform files, and calculate backazimuth and horizontal slowness.', formatter_class=argparse.ArgumentDefaultsHelpFormatter)
+def xcorr_all_terminal():
+    parser = argparse.ArgumentParser(description='Use cross-correlation to find delays between waveform files, and calculate backazimuth and horizontal slowness.', formatter_class=argparse.HelpFormatter)
     parser.add_argument('files', nargs='+', help='List of data files to process (wildcards are allowed)')
     
     parser.add_argument('-o', '--output_file', nargs = 1, default=None, help='Output file to write')
-    parser.add_argument('-i', '--include_IDs', default='', help='Station IDs to include in processing (default all)')
-    parser.add_argument('-x', '--exclude_IDs', default='', help='Station IDs to exclude from processing (default none)')
-    parser.add_argument('-1', '--t1', default='1970-01-01', help='Time to start processing data; default beginning of data')
-    parser.add_argument('-2', '--t2', default='9999-12-31', help='Time to stop processing data; default end of data')
-    parser.add_argument('-L', '--freq_low', default=5, help='Low corner frequency (default 5)')
-    parser.add_argument('-H', '--freq_high', default=40, help='High corner frequency (default 40)')
+    parser.add_argument('-i', '--include_IDs', help='Station IDs to include in processing (default: all)')
+    parser.add_argument('-x', '--exclude_IDs', help='Station IDs to exclude from processing (default: none)')
+    parser.add_argument('-1', '--t1', default='1970-01-01', help='Time to start processing data (default: beginning of data)')
+    parser.add_argument('-2', '--t2', default='9999-12-31', help='Time to stop processing data (default: end of data)')
+    parser.add_argument('-L', '--freq_low', default=5, help='Low corner frequency (default: 5 Hz)', type = float)
+    parser.add_argument('-H', '--freq_high', default=40, help='High corner frequency (default: 40 Hz)', type = float)
+    parser.add_argument('-w', '--window_length_seconds', default=5, help='Length of time windows to cross-correlate (default: 5 seconds)', type = float)
+    parser.add_argument('-p', '--overlap', default=0, help='Window overlap fraction; 0 <= overlap < 1 (default: 0)', type = float)
+    parser.add_argument('-u', '--upsample_ratio', default=1, help='Ratio by which waveform data should be upsampled to improve time resolution (default: 1 for no upsampling)', type = float)
+    parser.add_argument('-q', '--quiet', action = 'store_true')
     args = parser.parse_args()
-    
-    print(args.files)
+
+    if args.include_IDs is not None:
+        include_IDs = args.include_IDs.split(',')
+    else:
+        include_IDs = None
+    if args.exclude_IDs is not None:
+        exclude_IDs = args.exclude_IDs.split(',')
+    else:
+        exclude_IDs = None
+
     if args.output_file is None:
         raise(Exception('output_file is a required input'))
-    xcorr_df = xcorr_all(args.files, t1 = args.t1, t2 = args.t2)
-    xcorr_df.to_csv(args.output_file, sep = ',', index = False)
+
[…]
