+++ setup.py
@@ -100,7 +100,8 @@ ENTRY_POINTS = {
         'gem_cat = gemlog.gem_cat:main',
         'change_mseed_times = gemlog.change_mseed_times:main',
         'verify_huddle_test = gemlog.huddle_test:main',
-        'waveform_xc = gemlog.xcorr:main'
+        'waveform_xc = gemlog.xcorr:xcorr_all_terminal',
+        'calculate_direction = gemlog.xcorr:invert_for_slowness_terminal'
     ]
 }
 
