### 2023-03-21 9ae8da9
added tests for xcorr code using synthetic data
