### 2023-03-21 dffdd6a
renamed terminal commands, and bug fix
