### 2023-03-21 675dc7b
added new waveform cross-correlation terminal command
