+++ gemlog/version.py
@@ -1,4 +1,5 @@
-__version__ = '1.6.8' # added command-line tools to cross-correlate data and invert time lags for slowness and backazimuth 
+__version__ = '1.6.9' # added command-line tools to cross-correlate data and invert time lags for slowness and backazimuth, fixed for pypi 
+#__version__ = '1.6.8' # added command-line tools to cross-correlate data and invert time lags for slowness and backazimuth 
 #__version__ = '1.6.7' # bug fix in verify_huddle_test 
 #__version__ = '1.6.6' # version described by JOSS paper 
 #__version__ = '1.6.5' # bug fixes and other minor improvements 
@@ -60,4 +61,4 @@ __version__ = '1.6.8' # added command-line tools to cross-correlate data and inv
 #__version__  = '0.2.2' # automated github tests, setup.py improvements, and modelst speed-up
 #__version__  = '0.2.1' # 
 #__version__ = '0.0.5' # added new functions to make network map from gps data and rename mseeds from serial_number.channel to network.station.location.channel codes
-## List of old versions is not comprehensive#####################################################
+## List of old versions is not comprehensive######################################################
