### 2023-03-22 514d859
added examples to waveform_calc_lags help message
