### 2023-03-28 864dc6a
started new terminal command to process gps data
