### 2023-03-29 f6114e6
wrote terminal command to create inventory files
