+++ gemlog/gem_network.py
@@ -1,6 +1,6 @@
 import numpy as np
 import pandas as pd
-import glob, obspy, os, warnings, gemlog, sys
+import glob, obspy, os, warnings, gemlog, sys, argparse
 #from obspy.clients.nrl import NRL
 #from contextlib import contextmanager,redirect_stderr,redirect_stdout
 #from os import devnull
@@ -170,8 +170,12 @@ def _get_station_info(station_info):
                 raise Exception('invalid station_info file; must have 4 or 5 columns or valid header')
         ## if any keys are capitalized or abbreviated wrong, correct them
         station_info = _fix_station_info_keys(station_info)
-    elif (type(station_info) is not pd.DataFrame) or any([key not in station_info.keys() for key in required_keys]):
+    elif (type(station_info) is not pd.DataFrame):
         raise Exception('invalid station_info input')
+    else:
+        #station_info = _fix_station_info_keys(station_info)
+        if any([key not in station_info.keys() for key in required_keys]):
+            raise Exception('invalid station_info input')
     # if location and network fields are blank in file, they are interpreted as NaN and must be
     # turned back into blank
     if 'elevation' not in station_info.keys():
@@ -335,7 +339,7 @@ def rename_files(infile_pattern, station_info, output_dir, output_format = 'msee
         outfile_pattern = _fix_file_name_digits(outfile_pattern)
         output_file = outfile_pattern.format(**trace_info_dict)
         st.write(output_dir + '/' + output_file, format = output_format)
-        print(str(i) + ' of ' + str(len(infiles)) + ': ' + infile + ', ' + output_file)
+        print(str(i+1) + ' of ' + str(len(infiles)) + ': ' + infile + ', ' + output_file)
     return station_info
 
 def merge_files_day(infile_path, infile_pattern = '*', outfile_dir = 'merge_file_output'):
@@ -404,17 +408,25 @@ def read_gps(gps_dir_pattern, SN):
     pandas.DataFrame containing columns year, date (day of year), lat, lon (all floats), and column
     t (obspy.UTCDateTime).
     """
-    gpsDirList = sorted(glob.glob(gps_dir_pattern))
+    if type(gps_dir_pattern) is str:
+        gpsDirList = sorted(glob.glob(gps_dir_pattern))
+    elif type(gps_dir_pattern) is list:
+        gpsDirList = gps_dir_pattern
+    else:
+        raise Exception('invalid gps_dir_pattern type; must be list or str')
     gpsTable = pd.DataFrame(columns=['year', 'date', 'lat', 'lon', 't'])
     for gpsDir in gpsDirList:
-        fnList = sorted(glob.glob(gpsDir + '/' + SN + '*'))
+        if os.path.isdir(gpsDir):
+            fnList = sorted(glob.glob(gpsDir + '/' + SN + '*'))
+        else:
+            fnList = [gpsDir] # in case the user provides the gps filename directly
         if len(fnList) > 0: # if any gps files matching SN are found, read and append the last
             gpsTable = pd.concat([gpsTable,
                                   pd.read_csv(fnList[-1])], ignore_index=True)
     return gpsTable
 ReadLoggerGPS = read_gps # alias; v1.0.0
 
-def summarize_gps(gps_dir_pattern, station_info = None, output_file = None):
+def summarize_gps(gps_dir_pattern, station_info = None, output_file = None, t1 = None, t2 = None, include_SN = None, exclude_SN = None):
     """
     Read up-to-date GPS data from all Gems in a project, estimate their locations using a robust
     trimmed-mean method.
@@ -431,6 +443,18 @@ def summarize_gps(gps_dir_pattern, station_info = None, output_file = None):
     output_file : str
[…]
