### 2023-03-29 64530bd
added integration test for terminal inventory command
