### 2023-03-29 fbab1bd
new command-line tool to make sensor network info, including a stationXML file
