### 2023-04-18 #5256 @ajakef
Thanks for the great detailed review, @hemmelig! I'll work on addressing your comments. To answer one easy question, contributor TLSatterwhite is Tamara Beschorner, the third author on this paper.

@kthyng, should I wait for both reviews to finish before pushing changes? Or make changes as soon as I can?

Also, thanks for agreeing to review this, @thelenwes; I know it's a time commitment.
