### 2023-04-18 c6a1245
replaced % formatted strings with fstrings
