### 2023-04-18 ajakef/gemlog#42 @hemmelig
JOSS Review: Example of suggested code formatting changes
This is related to some code-formatting feedback as part of the review of `gemlog` for JOSS, over at  https://github.com/openjournals/joss-reviews/issues/5256#issuecomment-1489128525.

**Describe the suggested changes**
- Organise import section. Motivation here is to minimise unnecessary imports and use preferred standard library packages
- Move to `pathlib` over `os.path`. I don't think `os.path` can handle hardcoded `/` path delimiters when package running on Windows. `setup.py` indicates the package is intended to be OS independent.
- Apply Black auto-formatting, with line length set to 100 characters.

 - Operating System: macOS 12.5.1
 - Python version: 3.8
 - gemlog version: 1.7.0
