### 2023-04-25 d09a010
Create CODE_OF_CONDUCT.md
