### 2023-04-25 87cfe00
progress switching from os to pathlib library
