+++ gemlog/gem_cat.py
@@ -3,7 +3,7 @@ import pdb
 import warnings
 import numpy as np
 from numpy import NaN, Inf
-import os, glob, csv, time, scipy
+import os, glob, csv, time, scipy, pathlib
 import pandas as pd
 with warnings.catch_warnings():
     warnings.simplefilter("ignore")
@@ -29,12 +29,13 @@ def gem_cat(input_dir, output_dir, ext = '', cat_all = False):
     -------
     None; file output only
     """
-    if not os.path.isdir(input_dir):
+    if not pathlib.Path.is_dir(input_dir):
         raise Exception('Input path ' + input_dir + ' does not exist')
     
-    if not os.path.isdir(output_dir):
+    if not pathlib.Path.is_dir(output_dir):
         try:
-            os.makedirs(output_dir)
+            #os.makedirs(output_dir)
+            pathlib.Path(output_dir).mkdir(parents = True, exist_ok = True)
         except:
             raise Exception('Output path ' + output_dir + ' does not exist and cannot be created')
         else:
@@ -110,9 +111,11 @@ def gem_cat(input_dir, output_dir, ext = '', cat_all = False):
 #if True:
 def append_file(infile, outfile, prev_infile):
     # ensure that the output path exists
-    outpath = os.path.dirname(outfile)
-    if not os.path.exists(outpath):
-        os.makedirs(outpath)
+    #outpath = os.path.dirname(outfile)
+    outpath = pathlib.Path(outfile).parent
+    if not pathlib.Path.is_file(outpath):
+        #os.makedirs(outpath)
+        pathlib.Path(output).mkdir(parents = True, exist_ok = True)
         
     header = pd.read_csv(infile, delimiter = ',', nrows=1, dtype = 'str', names=['line']).line[0]
     format = header[7:]
