### 2023-04-25 ec6d3df
finished refactor to use pathlib instead of os
