+++ gemlog/huddle_test.py
@@ -4,7 +4,7 @@ import numpy as np
 import scipy.signal
 import matplotlib.pyplot as plt
 import matplotlib.dates as mdates
-import os, glob, getopt, sys, shutil, pathlib
+import glob, getopt, sys, shutil, pathlib
 import obspy, gemlog
 import time
 import datetime
@@ -54,7 +54,7 @@ def _min_starttime(SN_list, path):
     t1 = []
     t2 = []
     for SN in SN_list:
-        stream = obspy.read(path +'/mseed/*..' + SN + '..HDF.mseed')
+        stream = obspy.read(path / 'mseed' / ('*..' + SN + '..HDF.mseed'))
         trace = stream[0]
         t1.append(trace.starttime)
         t2.append(trace.endtime)
@@ -181,22 +181,19 @@ def verify_huddle_test(path, SN_list = [], SN_to_exclude = [], individual_only =
     """
     #%%
     if False: ## set default input values in development; set to True if running the code line-by-line
-        if os.getlogin() == 'jake':
-            path = '/home/jake/Work/gemlog_python/demo_QC'
-        else:
-            print('unknown user, need to define path')
+        #path = '/home/jake/Work/gemlog_python/demo_QC'
         SN_list = []
         SN_to_exclude = []
         individual_only = False
         run_crosscorrelation_checks = False
         generate_report = True
-        
+
+    path = pathlib.Path(path)
     ## Create folder for figures
-    figure_path = os.path.join(path, "figures")
-    file_exists = pathlib.Path.is_dir(figure_path)
+    figure_path = path / "figures"
+    file_exists = figure_path.is_dir()
     if file_exists == False:    
-        #os.mkdir(figure_path)
-        pathlib.Path(figure_path).mkdir(parents = True, exist_ok = True)
+        figure_path.mkdir(parents = True, exist_ok = True)
     else:
         pass
     
@@ -209,18 +206,18 @@ def verify_huddle_test(path, SN_list = [], SN_to_exclude = [], individual_only =
     ## Huddle test performance requirements:
     #### >=3 loggers must have barbs facing each other and all within 15 cm, in a turbulence-suppressed semi-enclosed space, sitting on a shared hard surface on top of padding, with good GPS signal, in a site that is not next to a continuous noise source (duty cycle < 80%). Loggers should all start and stop acquisition within 1 minute of each other, and run for at least one week.
 
-    metadata_path = path + '/metadata'
-    gps_path = path + '/gps'
-    mseed_path = path + '/mseed'
+    metadata_path = path / 'metadata'
+    gps_path = path / 'gps'
+    mseed_path = path / 'mseed'
     for test_path in [metadata_path, gps_path, mseed_path]:
         try:
-            fn = os.listdir(test_path)
+            fn = [x.name for x in test_path.glob('*')]
         except:
-            raise(Exception(test_path + ' not found'))
+            raise(Exception(f'{test_path} not found'))
 
     ## identify the list of serial numbers to test, including user input to include/exclude
     if len(SN_list) == 0:
-        SN_list = unique([filename[-14:-11] for filename in glob.glob(path + '/gps/*')])
+        SN_list = unique([filename[-14:-11] for filename in glob.glob(str(path / 'gps' / '*'))])
     SN_list = [SN for SN in SN_list if SN not in SN_to_exclude]
     print('Identified serial numbers ' + str(SN_list))
    
@@ -280,7 +277,7 @@ def verify_huddle_test(path, SN_list = [], SN_to_exclude = [], individual_only =
     ## Individual Metadata tests:
[…]
