### 2023-04-25 eaef762
bug fix (unsupported pandas append method)
