+++ gemlog/core.py
@@ -33,6 +33,7 @@ class MissingRawFiles(Exception):
     pass
 
 
+
 def _breakpoint():
     if _debug: # skip if we aren't in debug mode
         pdb.set_trace()
@@ -139,16 +140,16 @@ def convert(rawpath = '.', convertedpath = 'converted', metadatapath = 'metadata
     ## bitweight: leave blank to use default (considering Gem version, config, and units). This is preferred when using a standard Gem (R_g = 470 ohms)
     ## make sure the raw directory exists and has real data
     if not pathlib.Path(rawpath).is_dir():
-        raise MissingRawFiles('Raw directory ' + rawpath + ' does not exist')
+        raise MissingRawFiles(f'Raw directory {rawpath} does not exist')
     if len(glob.glob(rawpath + '/FILE' +'[0-9]'*4 + '.???')) == 0:
-        raise MissingRawFiles('No data files found in directory ' + rawpath)
+        raise MissingRawFiles(f'No data files found in directory {rawpath}')
     try:
         SN = str(SN)
         int(SN) # make sure it's number-like
     except:
         raise Exception('Invalid serial number')
     if len(SN) != 3:
-        raise Exception('Invalid serial number; SN type is length-'+ str(len(SN)) +' ' + str(type(SN)) + ', not length-3 str')
+        raise Exception(f'Invalid serial number; SN type is length-{str(len(SN))} {str(type(SN))}, not length-3 str')
     
     ## make sure bitweight is a scalar
     if((type(nums) is int) or (type(nums) is float)):
@@ -187,9 +188,9 @@ def convert(rawpath = '.', convertedpath = 'converted', metadatapath = 'metadata
     ## Catch if rawpath doesn't contain any files from SN. This won't catch files ending in TXT.
     if len(nums) == 0:
         if corrupt_files_flag:
-            raise CorruptRawFile('No non-corrupt data files for SN "' + SN + '" found in raw directory ' + rawpath)
+            raise CorruptRawFile(f'No non-corrupt data files for SN "{SN}" found in raw directory {rawpath}')
         else:
-            raise MissingRawFiles('No data files for SN "' + SN + '" found in raw directory ' + rawpath)
+            raise MissingRawFiles(f'No data files for SN "{SN}" found in raw directory {rawpath}')
 
     ## start at the first file in 'nums'
     nums.sort()
@@ -242,7 +243,7 @@ def convert(rawpath = '.', convertedpath = 'converted', metadatapath = 'metadata
             try:
                 pathlib.Path(gpspath).mkdir(parents = True, exist_ok = True)
             except:
-                print('Failed to make directory ' + gpspath)
+                print(f'Failed to make directory {gpspath}')
                 sys.exit(2)
         gpsfile = _make_filename(gpspath, SN, 'gps')
 
@@ -252,7 +253,7 @@ def convert(rawpath = '.', convertedpath = 'converted', metadatapath = 'metadata
             try:
                 pathlib.Path(metadatapath).mkdir(parents = True, exist_ok = True)
             except:
-                print('Failed to make directory ' + metadatapath)
+                print(f'Failed to make directory {metadatapath}')
                 sys.exit(2)
         metadatafile = _make_filename(metadatapath, SN, 'metadata')
   
@@ -261,7 +262,7 @@ def convert(rawpath = '.', convertedpath = 'converted', metadatapath = 'metadata
         try:
             pathlib.Path(convertedpath).mkdir(parents = True, exist_ok = True)
         except:
-            print('Failed to make directory ' + convertedpath)
+            print(f'Failed to make directory {convertedpath}')
[…]
