### 2023-04-26 c0aa781
added command to display terminal commands
