+++ gemlog/terminal_commands.py
@@ -1,0 +1,9 @@
+## command to print all the gemlog terminal commands
+def main():
+    print()
+    print('gemconvert: convert raw data to standard format (probably the tool you need)')
+    print('gemconvert_single: convert a single raw data file to standard format (less common)')
+    print('gem_cat: concatenate raw files as needed so all raw files have GPS (uncommon)')
+    print('gem_change_mseed_times: change miniSEED file start/end time while leaving data the same')
+    print('gem_verify_huddle_test: run a suite of data quality checks on data from a huddle test')
+    print()
