+++ setup.py
@@ -97,8 +97,9 @@ ENTRY_POINTS = {
         'gemconvert = gemlog.gemconvert:main',
         'gemconvert_single = gemlog.gemconvert_single:main',
         'gem_cat = gemlog.gem_cat:main',
-        'change_mseed_times = gemlog.change_mseed_times:main',
-        'verify_huddle_test = gemlog.huddle_test:main'
+        'gem_change_mseed_times = gemlog.change_mseed_times:main',
+        'gem_verify_huddle_test = gemlog.huddle_test:main',
+        'gemlog = gemlog.terminal_commands:main'
     ]
 }
 
