### 2023-04-26 2bd37b3
bug fix in read_gps when reading from a list of possible files
