+++ gemlog/gem_cat.py
@@ -5,6 +5,7 @@ import numpy as np
 from numpy import NaN, Inf
 import glob, csv, time, scipy, pathlib
 import pandas as pd
+
 with warnings.catch_warnings():
     warnings.simplefilter("ignore")
 import obspy
@@ -12,71 +13,76 @@ import sys
 import shutil
 from gemlog.core import _read_single, EmptyRawFile, CorruptRawFile
 
-def gem_cat(input_dir, output_dir, ext = '', cat_all = False):
+
+def gem_cat(input_dir, output_dir, ext="", cat_all=False):
     """
     gem_cat
     Merge raw data files so that all contain GPS data.
-    
+
     Translated from R code originally by Danny Bowman.
-    
+
     Parameters
     ----------
     input_dir: raw gem directory
     output_dir: path for renumbered and concatenated files
     ext: extension of raw files to convert (normally the serial number; sometimes TXT for old Gems)
-    
+
     Returns
     -------
     None; file output only
     """
     if not pathlib.Path(input_dir).is_dir():
-        raise Exception('Input path ' + input_dir + ' does not exist')
-    
+        raise Exception("Input path " + input_dir + " does not exist")
+
     if not pathlib.Path(output_dir).is_dir():
         try:
-            pathlib.Path(output_dir).mkdir(parents = True, exist_ok = True)
+            pathlib.Path(output_dir).mkdir(parents=True, exist_ok=True)
         except:
-            raise Exception('Output path ' + output_dir + ' does not exist and cannot be created')
+            raise Exception("Output path " + output_dir + " does not exist and cannot be created")
         else:
-            print('Created output folder ' + output_dir)
-    gem_files = sorted(glob.glob(input_dir + '/' + 'FILE[0-9][0-9][0-9][0-9].' + ext + '*'))
-    
+            print("Created output folder " + output_dir)
+    gem_files = sorted(glob.glob(input_dir + "/" + "FILE[0-9][0-9][0-9][0-9]." + ext + "*"))
+
     has_gps = np.zeros(len(gem_files))
     counter = 0
-    out_file = [] ## note that out_file being empty acts as a flag saying that no files have been processed yet
+    out_file = (
+        []
+    )  ## note that out_file being empty acts as a flag saying that no files have been processed yet
     try:
-        SN = pd.read_csv(gem_files[0], delimiter = ',', skiprows = 4, nrows=1, dtype = 'str', names=['s', 'SN']).SN[0]
+        SN = pd.read_csv(
+            gem_files[0], delimiter=",", skiprows=4, nrows=1, dtype="str", names=["s", "SN"]
+        ).SN[0]
     except:
         raise EmptyRawFile(gem_files[0])
 
     for k in range(len(gem_files)):
-        #for k in range(2):        
-        print(str(k+1) + ' of ' + str(len(gem_files)) + ': ' + gem_files[k])
+        # for k in range(2):
+        print(str(k + 1) + " of " + str(len(gem_files)) + ": " + gem_files[k])
 
         if cat_all:
             if len(out_file) == 0:
-                out_file = output_dir + '/FILE9999.' + SN
+                out_file = output_dir + "/FILE9999." + SN
                 shutil.copy(gem_files[k], out_file)
             else:
-                append_file(gem_files[k], out_file, gem_files[k-1])
+                append_file(gem_files[k], out_file, gem_files[k - 1])
             continue
 
         try:
             ## here and elsewhere, delimiter='\t' is used as a dummy to read each line as a whole string (read_csv doesn't accept '\n')
-            lines = pd.read_csv(gem_files[k], delimiter = '\t', dtype = 'str', names = ['line'])
+            lines = pd.read_csv(gem_files[k], delimiter="\t", dtype="str", names=["line"])
         except:
[…]
