+++ gemlog/__init__.py
@@ -1,6 +1,15 @@
-from gemlog import parsers # cython file
+from gemlog import parsers  # cython file
 from gemlog.core import Convert, ReadGem, convert, read_gem, get_gem_specs
-from gemlog.gem_network import summarize_gps, read_gps, SummarizeAllGPS, make_gem_inventory, rename_files, merge_files_day, get_gem_response, deconvolve_gem_response
+from gemlog.gem_network import (
+    summarize_gps,
+    read_gps,
+    SummarizeAllGPS,
+    make_gem_inventory,
+    rename_files,
+    merge_files_day,
+    get_gem_response,
+    deconvolve_gem_response,
+)
 from gemlog.gemlog_aux import make_db, calc_channel_stats, gem_noise, ims_noise
 from gemlog.gem_cat import gem_cat
 from gemlog.huddle_test import verify_huddle_test
