+++ gemlog/core.py
@@ -4,6 +4,7 @@ import numpy as np
 from numpy import NaN, Inf
 import glob, csv, time, scipy, pathlib
 import pandas as pd
+
 with warnings.catch_warnings():
     warnings.simplefilter("ignore")
 import obspy
@@ -13,93 +14,121 @@ import matplotlib.pyplot as plt
 
 _debug = False
 
+
 class EmptyRawFile(Exception):
     """Raised when the input file is empty"""
+
     pass
+
+
 class CorruptRawFile(Exception):
     """Raised when the input file cannot be read for some reason"""
+
     pass
+
 
 class CorruptRawFileNoGPS(CorruptRawFile):
     """Raised when the input file does not contain any GPS data"""
+
     pass
+
 
 class CorruptRawFileInadequateGPS(CorruptRawFile):
     """Raised when the input file does not contain any GPS data"""
+
     pass
+
 
 class MissingRawFiles(Exception):
     """Raised when no input files are readable"""
+
     pass
 
 
-
 def _breakpoint():
-    if _debug: # skip if we aren't in debug mode
+    if _debug:  # skip if we aren't in debug mode
         pdb.set_trace()
-        
+
+
 #####################
-def convert(rawpath = '.', convertedpath = 'converted', metadatapath = 'metadata', \
-            metadatafile = '', gpspath = 'gps', gpsfile = '', t1 = -Inf, t2 = Inf, nums = NaN, \
-            SN = '', bitweight = NaN, units = 'Pa', time_adjustment = 0, blockdays = 1, \
-            file_length_hour = 24, station = '', network = '', location = '', output_format = 'MSEED'):
+def convert(
+    rawpath=".",
+    convertedpath="converted",
+    metadatapath="metadata",
+    metadatafile="",
+    gpspath="gps",
+    gpsfile="",
+    t1=-Inf,
+    t2=Inf,
+    nums=NaN,
+    SN="",
+    bitweight=NaN,
+    units="Pa",
+    time_adjustment=0,
+    blockdays=1,
+    file_length_hour=24,
+    station="",
+    network="",
+    location="",
+    output_format="MSEED",
+):
     """
     Read raw Gem files, interpolate them, and write output files in miniSEED or SAC format.
 
     Parameters
     ----------
-    rawpath : str, default '.' 
+    rawpath : str, default '.'
         Path of folder containing raw Gem data files to read.
 
     convertedpath : str, 'converted'
-        Path of folder where converted data files should be written. If this folder does 
+        Path of folder where converted data files should be written. If this folder does
         not exist, convert will try to create it.
-    
+
     metadatapath : str, 'metadata'
-        Path of folder where metadata files should be written. If this folder does 
+        Path of folder where metadata files should be written. If this folder does
         not exist, convert will try to create it.
-    
+
     metadatafile : str
         File name for metadata file. Default is of the form 'XXXmetadata_NNN.txt', where
         'XXX' is the Gem's serial number and 'NNN' is the file index (000 at first, incrementing by one for each subsequent calculation).
 
     gpspath : str, 'gps'
-        Path of folder where gps files should be written. If this folder does 
+        Path of folder where gps files should be written. If this folder does
         not exist, convert will try to create it.
-    
+
     gpsfile : str
         File name for gps file. Default is of the form 'XXXgps_NNN.txt', where
         'XXX' is the Gem's serial number and 'NNN' is the file index (000 at first, incrementing by one for each subsequent calculation).
-    
+
     t1 : float or obspy.UTCDateTime, default -np.inf
-        Start time for the conversion. If float, the number of seconds 
-        since the epoch (1970-01-01 00:00:00 UTC). 
[…]
