### 2023-06-02 e848524
more work on demo notebooks
