+++ demo/README.md
@@ -1,141 +1,5 @@
 # Gem Data Pre-Processing Workflow
 
-### Installing the gemlog software
-Follow the gemlog [installation procedure](https://github.com/ajakef/gemlog/blob/main/Installation.md).
+This sample dataset and notebook show a typical data conversion, inspection, and station metadata routine.
 
-### Getting Started
-First, download the zip file in the list of files above; this contains the inputs for this demonstration. Move the file to some convenient folder, unzip it, and cd into the project folder.
-
-Notice the structure of this project folder. It includes a “raw” folder where the gem data files go. You don't have to use the folder name “raw”--or even have a separate folder for the raw data files--but all the default settings assume you're doing it this way.
-```
-Project_Folder/
-|__station_info.txt (table of SEED codes vs Gem serial number)
-|__raw/ (raw data from all loggers, FROM THIS PROJECT ONLY)
-```
-
-This example includes six two-hour files from six data loggers.
-```
-$ ls raw
-FILE0010.096
-FILE0010.103
-FILE0010.109
-FILE0011.088
-FILE0011.106
-FILE0012.077
-```
-Gem data files have the Gem's serial number as the extension, meaning that we have data from Gems 077, 088, 096, 103, 106, and 109. Empty or nearly-empty data files (like those from quick tests between field campaigns) can cause problems in the data conversion. It's a good idea to start each field campaign with clean disks, and to inspect your raw files and remove bad files before converting them.
-
-The optional file `station info.txt` is used to assign [SEED codes](https://ds.iris.edu/ds/nodes/dmc/data/formats/seed/) (network, station, and location) to your output data. By default, your converted data will use the Gem serial number as the station name, and leave the network and location codes blank. If you want to assign different names, you need to include a comma-separated text file containing a table of serial numbers, network codes, station codes, and location codes. Because all Gem recordings are infrasound sampled at 100 Hz, they are automatically assigned the [channel code](https://ds.iris.edu/ds/nodes/dmc/data/formats/seed-channel-naming/) HDF.
-
-### Converting raw gem data
-Set your conda environment to whatever you set up during the installation using a terminal command like `conda activate gem` and run the data conversion using the terminal command `gemconvert`. This may take a while to run if you have a lot of data! `gemconvert` will add the following folders to your directory structure:
-```
-Project_Folder/
-|__station_info.txt (unchanged, and unused in this step)
-|__raw/ (unchanged)
-|__mseed/ (NEW: contains hour-long miniSEED waveform files)
-|__metadata/ (NEW: contains csv files with state-of-health data)
-|__gps/ (NEW: contains gps logs)
-|__gemconvert_logfile.txt (NEW: contains a record of the conversion process and any messages)
-```
-
-By now, all the useful information has been extracted from the raw files; there is no reason to work with them further except for possible debugging. For more info on the gemconvert options and capabilities (e.g., the ability to write SAC files and the text format TSPAIR), run `gemconvert -h`.
-
-##### Multiple conversion attempts
-If the conversion is run multiple times, `gemconvert` will overwrite pre-existing miniSEED files. However, gps and metadata files are tagged with a conversion number at the end of their file name, so they are not overwritten. For a given Gem serial number, the file with the highest conversion number was created most recently. `gemconvert_logfile.txt `is appended to on each conversion attempt, so it is also not overwritten.
-
-### Inspect the Data
-The following workflow can be used to read the mseed data and deconvolve the instrument response.
[…]
