+++ README.md
@@ -37,3 +37,10 @@ User contributions are welcome! If you want to improve performance or address ou
 * Check the demonstrations mentioned above to see if it addresses your issue.
 * Look through the issues in this github site to see if anyone else is having the same trouble you are.
 * If those don't work, email project lead Jake Anderson at jacobanderson152@boisestate.edu.
+
+# Statement of need
+The Gem Infrasound Logger [@Anderson2018] is an approach to infrasound recording where the sensor and data logger are built into a single cable-free package that is easy to conceal and permits arbitrary sensor network geometries. Additionally, it has a low cost and weight, runs for months on alkaline batteries, and has a simple, fast installation procedure. These characteristics make it a good choice for temporary infrasound campaigns. By contrast, most campaigns that do not use the Gem use analog infrasound sensors that, via long cables, connect to multichannel data loggers built to record seismometers. This approach yields high-quality data but has several disadvantages: seismic data loggers are expensive, and sensor cables constrain the sensor network's geometry, make a station prone to animal damage and vandalism due to their visibility and exposure, and account for a large share of a recording site's budget for weight, bulk, and setup time. These disadvantages are especially acute for temporary recording campaigns (as opposed to permanent installations), which account for a large share of infrasound research.
+
+Like many geophysical data loggers, the Gem writes data in a non-standard raw format intended to balance firmware simplicity and performance, human readability, and compact file size. Although it is a human-readable text format that an expert can read and understand on a line-by-line basis, data files consist of hundreds of thousands of lines with complicated formatting, meaning that reading it as a spreadsheet or data frame is impractical. Further, operations like clock drift corrections, data decompression, and conversion to standard file formats or classes must be performed to make the data accessible in standard visualization and analysis software. Users often need to convert data from 10 or more infrasound loggers spanning several weeks, meaning that thousands of files including billions of data points must be processed efficiently. The `gemlog` Python package is a cross-platform tool to facilitate data conversion, and is essential for all Gem infrasound logger users.
+
+Some other geophysical data loggers used to record infrasound, for example the [DiGOS DataCube-3](https://digos.eu/seismology/) and Reftek [RT-130](https://www.passcal.nmt.edu/content/instrumentation/dataloggers/3-channel-dataloggers/reftek-rt-130-datalogger), have non-standard raw data formats that must be converted to standard formats by software distributed by the manufacturer. Like `gemlog`, they conduct data conversion as a simple command line operation. However, to the authors' knowledge, none of those software packages (or the raw data formats they convert) are open-source.
