+++ paper/reviewer_response_hemmelig.md
@@ -1,0 +1,125 @@
+I’ve gone through the JOSS checklist and, bar where indicated below, everything looks in order to me. The manuscript set out the motivations for—and capabilities of—the gemlog package in a clear manner.
+
+On a purely stylish note, would it be possible to highlight the package name and command line utility names with inline code formatting in the manuscript?
+*Done.*
+
+The reference list seems thorough, though my knowledge of the infrasound field is somewhat limited. The Data Report (line 115 of the original proof, Dannemann Dugick and Bowman) should probably read “TurboWave I and II…” as opposed to “TurboWave i and II”.
+*Done.*
+
+You might indicate somewhere (probably in the documentation, rather than setting it in stone in the JOSS paper) a rough definition of ‘low-cost’, based on the prices of the components used.
+*I agree that setting it in stone is unwise as it tends to change. I added a link to the Gem website where info on purchasing can be found. Like most (all?) geophysical instrument manufacturers that I know of, my prices and availability are dynamic enough that I currently only provide prices by request via quotes, though I do intend to begin posting estimated prices publicly on the website this year.*
+
+General comments
+Who is TLSatterwhite? They have made some contributions to the code in the repository. Besides that, the contribution and authorship looks good to me.
+*TLSatterwhite is the third author.*
+
+You might consider releasing the code and assigning it its own DOI using something like Zenodo.
+
+
+Per the GitHub Community Standards tab (under Insights), the repository is missing a few items:
+
+    Code of conduct
+    Contributing
+    Security policy
+    Issue templates
+    Pull request template
+
+The repository readme does include (good + clear) information about contribution guidelines, so you might consider putting these into the corresponding file as well just for the sake of completeness. I don't think you need all of the above, it was just a [sic]
+
+*Added a CODE_OF_CONDUCT.md file (the standard Contributor's Covenant) and, as suggested, moved the contributing info from the README to a dedicated CONTRIBUTING.md. Given the small scale of the project, the limited ability/intent to issue patches for old versions, and the small number of issues and PRs, we think that including a security policy, issue templates, and PR template is more formal than is needed or helpful in this case.*
+
+
+You might consider incorporating the documentation (e.g. for installation, for the demos, etc) into the existing readthedocs. You could then add some of the information from the paper regarding the problem the software is designed for/target audience to the landing page of the readthedocs site.
+*This is a good idea that I will implement in a future version.*
+
+
+How long does it take to convert, say, a 1 month campaign, per station? Just a ballpark figure would be useful (minutes, hours, days?). I get a rough estimate on the order of 10 minutes per station.
+*We added this info to the README.md and to the paper (on the order of 10 seconds per station-day).*
+
+Installation
+The installation process was smooth and clearly documented. I was able to install gemlog both directly from the Python Package Index (using pip install gemlog) and also downloading the source code and installing from there (pip install . from the source directory). The authors uses a GitHub Action to build and release the package to the Python Package Index.
+
+Tests
+Again, the authors use a GitHub Actions to automatically test and lint any changes made and pushed to either a branch of the main repo, or a fork. However, in order to run these tests locally on my own machine, I had to install pytest (and also had to clone the source repo). You might consider adding some information to the documentation detailing what someone might need to do to test any changes made locally, without having to push changes to a remote repo.
+*Done; we added more explicit info on pytest to Installation.md and CONTRIBUTING.md.*
+
+Tests all ran locally without any issues, though there were a number of warnings that seemed to be about the implementation of the tests themselves.
+
+
+Demos
[…]
