### 2023-06-02 #5256 @ajakef
First, thank you both @hemmelig and @thelenwes for your  helpful reviews. I understand that reviewing someone else's software is a significant effort and I appreciate the time you put into it. Thanks also @bmcfee for overseeing this manuscript.

I've updated the `paper` branch with edits reflecting the reviewer comments (attached).
[reviewer_response_thelenwes.md](https://github.com/openjournals/joss-reviews/files/11639212/reviewer_response_thelenwes.md)
[reviewer_response_hemmelig.md](https://github.com/openjournals/joss-reviews/files/11639213/reviewer_response_hemmelig.md)

Regarding the checklists: I see that @hemmelig's checklist has the documentation "statement of need" box and the "references" box unchecked. I have made the change to the references noted in the paper, and I also added a "statement of need" to the main README.md. @bmcfee, does this look ok, and are there additional things I need to do?
