### 2023-06-05 #5256 @bmcfee
Thanks @ajakef !  (And sorry for the comment spam with the bot.)

Please check the above comment https://github.com/openjournals/joss-reviews/issues/5256#issuecomment-1576823589 for a final checklist of (author) tasks to complete, and I'll get to work on the editor tasks.
