+++ paper/paper.md
@@ -55,7 +55,7 @@ The Gem Infrasound Logger (including `gemlog`) has been included in independent 
 # Features
 [gemlog](https://github.com/ajakef/gemlog) (GPL-3 license) is a Python library that includes both terminal commands and Python functions. It is installable from [PyPI](https://pypi.org/project/gemlog/) and documented in terminal command help pages, Python function docstrings, and on [Read the Docs](https://gemlog.readthedocs.io/).
 
-- Data conversion: Terminal commands `gemconvert` and `gemconvert_single` convert sets of raw files into standard data formats (typically miniSEED). Because infrasound analyses require sample timing to be approximately millisecond-precise, an essential part of this process is correcting clock drift using accurate times provided infrequently by the Gem's on-board GPS. Contiguous blocks of raw data are converted into contiguous blocks of output data, and the software identifies breaks in recording and includes the same breaks in the output. On a typical laptop, conversion may take on the order of 10 wall-clock seconds per day of data for one station (or, a unitless ratio of approximately $10^-4$ between conversion time and data duration).
+- Data conversion: Terminal commands `gemconvert` and `gemconvert_single` convert sets of raw files into standard data formats (typically miniSEED). Because infrasound analyses require sample timing to be approximately millisecond-precise, an essential part of this process is correcting clock drift using accurate times provided infrequently by the Gem's on-board GPS. Contiguous blocks of raw data are converted into contiguous blocks of output data, and the software identifies breaks in recording and includes the same breaks in the output. On a typical laptop, conversion may take on the order of 10 wall-clock seconds per day of data for one station (or, a unitless ratio of approximately $10^{-4}$ between conversion time and data duration).
 
 - Instrument testing: Terminal command `verify_huddle_test` is used to automatically examine waveform, state-of-health, and GPS data from several instruments recording in the same place at the same time (termed a "huddle test" in seismology/infrasound), and verify that all instruments are working properly.
 
