+++ paper/paper.bib
@@ -78,7 +78,7 @@
 }
 @article{Averbuch2022,
   title={Evidence for Short Temporal Atmospheric Variations Observed by Infrasonic Signals: 1. The Troposphere},
-  author={Averbuch, G and Ronac-Giannone, M and Arrowsmith, S and Anderson, JF},
+  author={Averbuch, Gil and Ronac-Giannone, Miro and Arrowsmith, Stephen and Anderson, Jacob F},
   journal={Earth and Space Science},
   volume={9},
   number={3},
@@ -89,7 +89,7 @@
 }
 
 @article{Anderson_inprep,
-author = {Anderson, J.F. and Johnson, J.B. and Mikesell, T.D. and Liberty, L.M.},
+author = {Anderson, Jacob F. and Johnson, Jeffrey B. and Mikesell, Thomas D. and Liberty, Lee M.},
 title = {Remote detection of earthquake ground shaking via high-resolution large-N infrasound beamforming},
 year = {in prep.}
 }
@@ -97,7 +97,7 @@ year = {in prep.}
 
 @inproceedings{Gauvain2021,
   title={Variability of acoustic properties among whitewater features},
-  author={Gauvain, Scott and Tatum, Taylor and Bosa, Ashley and Satterwhite, Tamara and Anderson, Jacob},
+  author={Gauvain, Scott J and Tatum, Taylor and Bosa, Ashley and Satterwhite, Tamara and Anderson, Jacob F},
   booktitle={AGU Fall Meeting Abstracts},
   volume={2021},
   pages={S55C--0162},
@@ -143,7 +143,7 @@ year = {in prep.}
 }
 @inproceedings{Bosa2021,
   title={Dynamics of rain-triggered lahars inferred from infrasound array and time-lapse camera correlation at {V}olcan de {F}uego, {G}uatemala.},
-  author={Bosa, Ashley and Pineda, Armando and Mock, Jerry and Roca, Amilcar and Johnson, Jeffrey and Bejar, Gustavo and Waite, Gregory and Gauvain, Scott and Bartel, Beth},
+  author={Bosa, Ashley and Pineda, Armando and Mock, Jerry and Roca, Amilcar and Johnson, Jeffrey B and Bejar, Gustavo and Waite, Gregory and Gauvain, Scott J and Bartel, Beth},
   booktitle={AGU Fall Meeting Abstracts},
   volume={2021},
   pages={V25D--0150},
@@ -151,7 +151,7 @@ year = {in prep.}
 }
 @inproceedings{Mock2020,
   title={Topographic Effects of Infrasound Wave Propagation},
-  author={Mock, Jerry and Anderson, Jacob and Johnson, Jeffrey Bruce and Rosenblatt, Bryan and Gauvain, Scott},
+  author={Mock, Jerry and Anderson, Jacob F and Johnson, Jeffrey B and Rosenblatt, Bryan and Gauvain, Scott J},
   booktitle={AGU Fall Meeting Abstracts},
   volume={2020},
   pages={S001--0002},
