### 2023-06-07 #5256 @ajakef
Thanks @bmcfee. I fixed the 10^-4 issue.

Question: parenthetical citations of both of my own papers include my first name. No other citations do that. Is this a normal part of JOSS's format? I'd prefer that it not do that because it looks like I'm highlighting my own work and not others'.

I went through the author checklist. Am I supposed to be able to check those boxes? It mostly doesn't let me click them, so here it is here.
Done: Double check authors and affiliations (including ORCIDs)

Done (1.6.5.0): Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.

Done (10.5281/zenodo.8015143): Archive the release on Zenodo/figshare/etc and post the DOI here.

Done: Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.

Done: Make sure that the license listed for the archive is the same as the software license.
