### 2023-06-08 #5256 @bmcfee
Thanks @ajakef - sorry for the confusion on the checklist, but I can check these off for you as they're completed.

> I fixed the 10^-4 issue.

:+1:

> Question: parenthetical citations of both of my own papers include my first name. No other citations do that. Is this a normal part of JOSS's format? I'd prefer that it not do that because it looks like I'm highlighting my own work and not others'.

This is strange.  I wonder if it's getting confused because your bibtex has both `Anderson, Jacob F` and `Anderson, J.F.`:
https://github.com/ajakef/gemlog/blob/cdbd83c415aaba2bb7e45a96e258b619c017fcd3/paper/paper.bib#L3

https://github.com/ajakef/gemlog/blob/cdbd83c415aaba2bb7e45a96e258b619c017fcd3/paper/paper.bib#L92

Probably best to standardize them first in the bibtex source and see if that resolves things.
