### 2023-06-12 #5256 @bmcfee
@ajakef :wave: had a chance to look into the bib entries yet?  I think this is the only thing holding us up at this point.  No rush, but let me know when it's ready for a final pass.
