### 2023-06-12 d708c2c
minor changes to .bib file
