+++ paper/paper.bib
@@ -89,7 +89,7 @@
 }
 
 @article{Anderson_inprep,
-author = {Anderson, Jacob F. and Johnson, Jeffrey B. and Mikesell, Thomas D. and Liberty, Lee M.},
+author = {Anderson, Jacob F and Johnson, Jeffrey B and Mikesell, Thomas D and Liberty, Lee M},
 title = {Remote detection of earthquake ground shaking via high-resolution large-N infrasound beamforming},
 year = {in prep.}
 }
