### 2023-06-12 9d01a14
added and updated references
