+++ paper/paper.bib
@@ -88,12 +88,29 @@
   doi = {10.1029/2021ea002036}  
 }
 
-@article{Anderson_inprep,
+@article{Anderson_submitted,
 author = {Anderson, Jacob F and Johnson, Jeffrey B and Mikesell, Thomas D and Liberty, Lee M},
 title = {Remote detection of earthquake ground shaking via high-resolution large-N infrasound beamforming},
-year = {in prep.}
+year = {submitted}
 }
 
+@article{Tatum_submitted,
+author = {Tatum, Taylor A and Anderson, Jacob F and Ronan, Timothy J}
+title = {Whitewater Sound Dependence on Discharge and Wave Configuration at an Adjustable Wave Feature},
+year = {submitted}
+}
+
+@article{Scamfer_submitted,
+author = {Scamfer, Logan T and Anderson, Jacob F},
+title = {Exploring background noise with a large-N infrasound array: waterfalls, thunderstorms, and earthquakes},
+year = {submitted}
+}
+
+@article{Gauvain_inprep,
+author = {Gauvain, Scott J and Anderson, Jacob F and Yager, Elowyn M and McNamara, James P},
+title = {Effects of Discharge and Morphology on Fluvial Sound},
+year = {submitted}
+}
 
 @inproceedings{Gauvain2021,
   title={Variability of acoustic properties among whitewater features},
