+++ paper/paper.bib
@@ -95,7 +95,7 @@ year = {submitted}
 }
 
 @article{Tatum_submitted,
-author = {Tatum, Taylor A and Anderson, Jacob F and Ronan, Timothy J}
+author = {Tatum, Taylor A and Anderson, Jacob F and Ronan, Timothy J},
 title = {Whitewater Sound Dependence on Discharge and Wave Configuration at an Adjustable Wave Feature},
 year = {submitted}
 }
