### 2023-06-12 b197d95
minor tweaks to paper and bibliography
