+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{Anderson2018,
-  title={The Gem infrasound logger and custom-built instrumentation},
+  title={The {Gem} infrasound logger and custom-built instrumentation},
   author={Anderson, Jacob F and Johnson, Jeffrey B and Bowman, Daniel C and Ronan, Timothy J},
   journal={Seismological Research Letters},
   volume={89},
@@ -23,7 +23,7 @@
 }
 
 @article{Krishnamoorthy2020,
-  title={Infrasound Measurements in the Arctic on a Long-Duration Balloon},
+  title={Infrasound Measurements in the {Arctic} on a Long-Duration Balloon},
   author={Krishnamoorthy, Siddharth and Bowman, Daniel C and Komjathy, Attila and Pauken, Michael T and Cutts, James A and Carlo, Marine De and Ardhuin, Fabrice and Le Pichon, Alexis},
   year={2020},
   publisher={Pasadena, CA: Jet Propulsion Laboratory, National Aeronautics and Space Administration},
@@ -90,7 +90,7 @@
 
 @article{Anderson_submitted,
 author = {Anderson, Jacob F and Johnson, Jeffrey B and Mikesell, Thomas D and Liberty, Lee M},
-title = {Remote detection of earthquake ground shaking via high-resolution large-N infrasound beamforming},
+title = {Remote detection of earthquake ground shaking via high-resolution large-{N} infrasound beamforming},
 year = {submitted}
 }
 
@@ -102,14 +102,14 @@ year = {submitted}
 
 @article{Scamfer_submitted,
 author = {Scamfer, Logan T and Anderson, Jacob F},
-title = {Exploring background noise with a large-N infrasound array: waterfalls, thunderstorms, and earthquakes},
+title = {Exploring background noise with a large-{N} infrasound array: waterfalls, thunderstorms, and earthquakes},
 year = {submitted}
 }
 
 @article{Gauvain_inprep,
 author = {Gauvain, Scott J and Anderson, Jacob F and Yager, Elowyn M and McNamara, James P},
 title = {Effects of Discharge and Morphology on Fluvial Sound},
-year = {submitted}
+year = {in prep.}
 }
 
 @inproceedings{Gauvain2021,
