### 2023-06-12 #5256 @ajakef
Thanks @bmcfee, your citation fix worked. I also made minor formatting changes and added a couple references that didn't exist at the time of submission; hope that's ok. From my point of view it's ready to proceed.
