### 2023-06-13 #5256 @bmcfee
@ajakef Yes - I think in that case, it's arguably more important to cite the DOI of the version that exists at the time of this submission, just to prevent confusion later on. (I expect the contents of the cited work won't change substantially upon review later on, but that's somewhat beside the point.)
