### 2023-06-14 b235656
added DOI for Tatum et al., 2023 (ESSOAR version)
