+++ paper/paper.bib
@@ -97,7 +97,9 @@ year = {submitted}
 @article{Tatum_submitted,
 author = {Tatum, Taylor A and Anderson, Jacob F and Ronan, Timothy J},
 title = {Whitewater Sound Dependence on Discharge and Wave Configuration at an Adjustable Wave Feature},
-year = {submitted}
+year = {2023},
+journal = {ESS Open Archive},
+doi = {10.22541/essoar.167397475.53275976/v1}
 }
 
 @article{Scamfer_submitted,
